make the idle screensaver a yard of ducks
Ctrl+Y already existed as a hidden shortcut that shows the screensaver
immediately, so wiring it up was really a matter of making the screensaver be
this. The simulation moves from scripts/ into ui/shared/_mayhem.py — scripts/ is
not importable from an installed yeaboi, so nothing in the app could have
reached it there. scripts/demo_duck_mayhem.py stays as a driver over the same
code, for the two things a recording needs and an idle screen does not: a
sprite scale higher than any real terminal would use, and a fixed run length.

build_screensaver hands off to it at 60x24 and up. Below that the crowd has
nowhere to go and it reads as a jam rather than as mayhem, so the existing
single-duck bands still cover the small end.

render() keeps one yard and advances it in place. build_screensaver is called
once per frame with a rising elapsed, so stepping incrementally is O(1) a frame
where recomputing from zero would get slower the longer it ran. It rebuilds when
the terminal is resized or when time runs backwards — backwards being the
ordinary case, since every session starts its clock again — and caps catch-up at
two seconds, or a session resumed after the machine slept would try to simulate
hours in one frame and hang the UI.

Drift and spin are halved. They were tuned for an eight-second clip that had to
earn attention; this runs until somebody touches a key.
